+++ index_sse2.h
@@ -1,6 +1,6 @@
 #ifndef CSIMDBYTE_INDEX_SSE2_H
 #define CSIMDBYTE_INDEX_SSE2_H
 
-int index_sse2(const char *s, int sz, const char *ss, int ssz);
+size_t index_sse2(const char *s, size_t sz, const char *ss, size_t ssz);
 
 #endif //CSIMDBYTE_INDEX_SSE2_H
